requirements.txt

beautifulsoup4==4.12.3

crewai==0.22.5
crewai-tools==0.0.15

Flask==3.0.3
Flask-Cors==4.0.1

groq==0.5.0

langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-groq==0.1.3
langchain-openai==0.0.5

openai==1.25.1


PyPDF2==3.0.1

python-dotenv==1.0.0

requests==2.31.0


urllib3==2.2.1
utils==1.0.2

python = 3.11
pydantic == 2.6.3
load-dotenv == 0.1.0
